tests(apis): add api client tests; update factories and conftest; correct functions and view

+++ helpers/functions.py
@@ -5,34 +5,45 @@ from accounts.models import Employee
 from apis.serializers import ClientDetailSerializer
 
 
+def check_uuid_value(uuid_value):
+    """Return error message or None."""
+    try:
+        uuid.UUID(str(uuid_value))
+    except ValueError:
+        return "Veuillez saisir un UUID valide."
+
+
 def add_locations(model_instance, location_data):
+    """Add location(s) to model."""
     for location_item in location_data:
         location, created = Location.objects.get_or_create(**location_item)
         model_instance.locations.add(location)
 
 
 def remove_locations(model_instance, location_id):
-    """Remove location from model if it belongs to it
-    and delete location from the application if it is not used by other clients or events.
     """
-    for location_item in model_instance.locations.all():
-        if str(location_item.location_id) == location_id:
-            model_instance.locations.remove(location_item.location_id)
-            locations = Location.objects.all()
-            for location in locations:
-                if location_id == str(location.location_id):
-                    if (
-                        location.client_locations.count() == 0
-                        and location.event_locations.count() == 0
-                    ):
-                        location.delete()
-
-
-def check_uuid_value(uuid_value):
-    try:
-        uuid.UUID(str(uuid_value))
-    except ValueError:
-        return "Veuillez saisir un UUID valide."
+    Remove location from model if exists and belongs to it
+    and delete location from the application if it is not used by other clients or events.
+    Return error message or None
+    """
+    location_uuid = check_uuid_value(location_id)
+    if not isinstance(location_uuid, str):
+        if Location.objects.filter(location_id=location_id).exists():
+            for location_item in model_instance.locations.all():
+                if str(location_item.location_id) == location_id:
+                    model_instance.locations.remove(location_item.location_id)
+                    locations = Location.objects.all()
+                    for location in locations:
+                        if location_id == str(location.location_id):
+                            if (
+                                location.client_locations.count() == 0
+                                and location.event_locations.count() == 0
+                            ):
+                                location.delete()
+        else:
+            return "Le lieu n'existe pas."
+    else:
+        return location_uuid
 
 
 def update_sales_contact(client_instance, updated_sales_contact_id):
